refactor: collapse act-local-ci into scripts and inline integration port
Area E (CI audit §5): move pure helpers from src/core/act-local-ci.ts into
scripts/act-local-ci.ts; drop core barrel re-exports; merge tests into
test/scripts/act-local-ci.test.ts. Inline python ephemeral-port in
integration.yml; delete scripts/integration-ephemeral-port.sh. Document
nix-run-if-missing.sh callers; record FC/IS exception in AGENTS.md; refresh
docs and bunfig coverage note.
