+++ .env.example
@@ -25,7 +25,7 @@
 
 # --- Local act (`bun run act`) — optional; see CONTRIBUTING.md ---
 # Omit ACT_RUNS_ON_LABEL to use the script default (aligned with this repo’s workflow `runs-on`). Set it only when simulating jobs whose label differs.
-# Omit ACT_RUNNER_IMAGE to use the script defaults (see resolveActRunnerImage in src/core/act-local-ci.ts). Set it only to pin or swap the container image for `act -P`.
+# Omit ACT_RUNNER_IMAGE to use the script defaults (see resolveActRunnerImage in scripts/act-local-ci.ts). Set it only to pin or swap the container image for `act -P`.
 # Do not set GITHUB_ACTIONS in .env unless you intentionally want to mimic GitHub Actions runner
 # behavior (e.g. smaller default image for dry-run); leave it unset for normal local `act` runs.
