+++ CONTRIBUTING.md
@@ -45,9 +45,9 @@ Without these tools installed, `scripts/nix-run-if-missing.sh` will use `nix run
 - **act standalone**: `brew install act` (or [other install options](https://github.com/nektos/act#installation))
 - **Nix**: With Nix installed, `act` is available from this flake on **Linux and macOS** (see `flake.nix` systems). The helper runs `nix run .#act` when `act` is not on your PATH (same pattern as `scripts/nix-run-if-missing.sh`). `nix develop` also puts `act` in PATH. Without Nix, use `gh act`, Homebrew `act`, or another install.
 
-The script prefers standalone `act` (PATH or `nix run .#act` when the flake supports your host), so you do not need `gh extension install nektos/gh-act` unless you rely on `gh act` alone. If neither `act` nor Nix is available, it uses `gh act` when that extension is installed. In code this is the **`direct`** backend (`bash` + `scripts/nix-run-if-missing.sh` + `act`) vs **`gh`** (`gh act`); see `ActBackend` in [`act-local-ci.ts`](src/core/act-local-ci.ts).
+The script prefers standalone `act` (PATH or `nix run .#act` when the flake supports your host), so you do not need `gh extension install nektos/gh-act` unless you rely on `gh act` alone. If neither `act` nor Nix is available, it uses `gh act` when that extension is installed. In code this is the **`direct`** backend (`bash` + `scripts/nix-run-if-missing.sh` + `act`) vs **`gh`** (`gh act`); see `ActBackend` in [`scripts/act-local-ci.ts`](scripts/act-local-ci.ts).
 
-This repo’s workflows use a specific `runs-on` label (see the YAML). [act](https://github.com/nektos/act) needs **`-P <runs-on>=<container image>`** where `<runs-on>` **matches** the workflow job’s label. [scripts/act-local-ci.ts](scripts/act-local-ci.ts) defaults the label to **`ubuntu-24.04`** ([`DEFAULT_ACT_RUNS_ON_LABEL`](src/core/act-local-ci.ts)); set **`ACT_RUNS_ON_LABEL`** if your workflows use something else. It resolves a **container image** via **`ACT_RUNNER_IMAGE`** (optional) or built-in defaults tuned for this repo’s current label. Same script writes a minimal `workflow_dispatch` JSON (`-e`, from `git remote origin` or `package.json` `repository`) to `.act-artifacts/workflow_dispatch.json` so `github.event.repository` exists for actions that need it, and **`--artifact-server-path .act-artifacts/`** so [`actions/upload-artifact`](https://github.com/actions/upload-artifact) (SBOM, gitleaks SARIF) gets a local [artifact server](https://github.com/nektos/act/issues/1929) instead of GitHub’s `ACTIONS_RUNTIME_TOKEN`. If you invoke **`act` by hand**, pass **`-P`** with the same label as `runs-on` and your chosen image.
+This repo’s workflows use a specific `runs-on` label (see the YAML). [act](https://github.com/nektos/act) needs **`-P <runs-on>=<container image>`** where `<runs-on>` **matches** the workflow job’s label. [scripts/act-local-ci.ts](scripts/act-local-ci.ts) defaults the label to **`ubuntu-24.04`** ([`DEFAULT_ACT_RUNS_ON_LABEL`](scripts/act-local-ci.ts)); set **`ACT_RUNS_ON_LABEL`** if your workflows use something else. It resolves a **container image** via **`ACT_RUNNER_IMAGE`** (optional) or built-in defaults tuned for this repo’s current label. Same script writes a minimal `workflow_dispatch` JSON (`-e`, from `git remote origin` or `package.json` `repository`) to `.act-artifacts/workflow_dispatch.json` so `github.event.repository` exists for actions that need it, and **`--artifact-server-path .act-artifacts/`** so [`actions/upload-artifact`](https://github.com/actions/upload-artifact) (SBOM, gitleaks SARIF) gets a local [artifact server](https://github.com/nektos/act/issues/1929) instead of GitHub’s `ACTIONS_RUNTIME_TOKEN`. If you invoke **`act` by hand**, pass **`-P`** with the same label as `runs-on` and your chosen image.
 
 **Docker/act gotchas:** The [check](.github/workflows/check.yml) job sets **`RUNNER_TOOL_CACHE`** and **`AGENT_TOOLSDIRECTORY`** under **`github.workspace/.cache`** so **`actions/setup-node`** does not try to create **`/opt/hostedtoolcache/...`** as an unprivileged user (fixes common **`EACCES`** in act). By default **`act-local-ci`** does **not** pass **`--artifact-server-port`** (act listens on its default, usually **34567**) so **[`actions/upload-artifact`](https://github.com/actions/upload-artifact)** and gitleaks SARIF work. **`--artifact-server-port=0`** breaks artifact uploads with current act (bad URL). If **`listen … :34567: address already in use`**, set **`ACT_ARTIFACT_SERVER_PORT`** to a free port (e.g. **`45678`**) or stop a stale **`act`**. **`ACT_ARTIFACT_SERVER_PORT=-`** omits the flag; optional **`ACT_ARTIFACT_SERVER_ADDR`**.
