+++ bunfig.toml
@@ -9,7 +9,7 @@ pathIgnorePatterns = ["test/integration/**"]
 coverage = true
 coverageReporter = ["text", "lcov"]
 coverageDir = "coverage"
-# act-local-ci.ts: CLI + process I/O; per-file coverage still misses most branches. Pure logic: src/core/act-local-ci.ts; program dry-run path: test/scripts/act-local-ci.test.ts.
+# act-local-ci.ts: CLI + process I/O; per-file coverage still misses most branches. Pure argv helpers: same file; program dry-run path: test/scripts/act-local-ci.test.ts.
 coveragePathIgnorePatterns = ["test/**", "**/core/errors.ts", "scripts/act-local-ci.ts"]
 coverageSkipTestFiles = true
 # Bun enforces per-file threshold on both % funcs and % lines. Floor was 0.68 (weakest line ~68%).
