+++ docs/CI.md
@@ -111,7 +111,7 @@ Separately, [update-flake-lock.yml](../.github/workflows/update-flake-lock.yml) 
 
 **Docker parity:** [check.yml](../.github/workflows/check.yml) sets **`RUNNER_TOOL_CACHE`** / **`AGENT_TOOLSDIRECTORY`** under the workspace so **`actions/setup-node`** (SBOM step) avoids **`EACCES`** on `/opt/hostedtoolcache` in act. [act-local-ci.ts](../scripts/act-local-ci.ts) leaves **`--artifact-server-port`** at act’s default unless you set **`ACT_ARTIFACT_SERVER_PORT`** / **`ACT_ARTIFACT_SERVER_ADDR`** (**`0`** is not a safe default—artifact uploads can fail). Details: [CONTRIBUTING.md](../CONTRIBUTING.md#run-ci-locally-check-job).
 
-**Integration + nested Docker:** On Linux, [act-local-ci.ts](../scripts/act-local-ci.ts) passes **`act --container-options`** with **`--group-add`** set from the host **`/var/run/docker.sock`** group id when that path exists, so integration jobs can run **`docker`** inside the act job container. Override with **`ACT_CONTAINER_OPTIONS`** or disable auto with **`ACT_SKIP_AUTO_DOCKER_GROUP_ADD=1`**. Rootless Docker or a nonstandard socket may need manual **`ACT_CONTAINER_OPTIONS`**. [integration.yml](../.github/workflows/integration.yml) caches llama artifacts under **`github.workspace/.cache/`** (stable vs **`runner.temp`** when debugging paths). The **llama-server-docker-start** action copies the GGUF into the container with **`docker cp`**, not a bind mount, so nested Docker (act) does not rely on host path alignment. Parallel llama jobs each bind an **ephemeral TCP port** on the runner (`scripts/integration-ephemeral-port.sh` via **`python3`**, as on GitHub-hosted Ubuntu) and use distinct **`container_name`** values so they do not collide on a single Docker daemon. Only when simulating **`integration.yml`** (not **`check`** / **`check-workflows`**) does [act-local-ci.ts](../scripts/act-local-ci.ts) pass **`--env AUTO_PR_ACT_LOCAL_CI=1`** so the happy-path integration test can use a longer timeout under CPU-bound [nektos/act](https://github.com/nektos/act).
+**Integration + nested Docker:** On Linux, [act-local-ci.ts](../scripts/act-local-ci.ts) passes **`act --container-options`** with **`--group-add`** set from the host **`/var/run/docker.sock`** group id when that path exists, so integration jobs can run **`docker`** inside the act job container. Override with **`ACT_CONTAINER_OPTIONS`** or disable auto with **`ACT_SKIP_AUTO_DOCKER_GROUP_ADD=1`**. Rootless Docker or a nonstandard socket may need manual **`ACT_CONTAINER_OPTIONS`**. [integration.yml](../.github/workflows/integration.yml) caches llama artifacts under **`github.workspace/.cache/`** (stable vs **`runner.temp`** when debugging paths). The **llama-server-docker-start** action copies the GGUF into the container with **`docker cp`**, not a bind mount, so nested Docker (act) does not rely on host path alignment. Parallel llama jobs each bind an **ephemeral TCP port** on the runner (inline **`python3`** one-liner in [integration.yml](../.github/workflows/integration.yml), as on GitHub-hosted Ubuntu) and use distinct **`container_name`** values so they do not collide on a single Docker daemon. Only when simulating **`integration.yml`** (not **`check`** / **`check-workflows`**) does [act-local-ci.ts](../scripts/act-local-ci.ts) pass **`--env AUTO_PR_ACT_LOCAL_CI=1`** so the happy-path integration test can use a longer timeout under CPU-bound [nektos/act](https://github.com/nektos/act).
 
 **GitHub Models job:** The **`integration-github-models`** job loads **`.env.ci`** with **`omit_llama_integration_env: true`** so llama-related keys and **`AUTO_PR_AI_OPENAI_COMPAT_URL`** from **`.env.ci`** are not applied (that job uses GitHub Models only).
