+++ docs/INTEGRATION.md
@@ -225,7 +225,7 @@ Replace `<SHA>` with the SHA from the `uses:` lines in [auto-pr.yml](../.github/
 - **Tag vs digest:** The template uses a **tag** on `FROM` (e.g. `ghcr.io/ggml-org/llama.cpp:server`) so [Dependabot](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#docker) can propose image updates. For a **stricter immutable** pin, use `image@sha256:…` on the same line (Dependabot behavior may differ from tag-only pins).
 - **Parser limits (workflow + tests):** Only the **first** `FROM` is used, after optional `--` flags (e.g. `--platform=…`). There is **no** support for backslash line continuation or for picking a later stage in a **multi-stage** file—keep this file **single-stage**, or ensure the image you need appears on the **first** `FROM`.
 - **Docker llama composite actions:** Start/stop are [`llama-server-docker-start`](../.github/actions/llama-server-docker-start) and [`llama-server-docker-stop`](../.github/actions/llama-server-docker-stop) (pinned in the reusable workflow). The start script uses **`docker cp`** to place the GGUF in the container (not a bind mount), so nested Docker (e.g. [act](https://github.com/nektos/act)) does not depend on host path alignment for `-v`. Default container name is **`auto-pr-llama`**; pass **`container_name`** when several jobs share one Docker host ([nektos/act](https://github.com/nektos/act) runs parallel jobs on a single machine). Assume **one** `llama-server` container per job unless you use distinct **`container_name`** and **host port** values. If you copy those composite actions into a custom workflow and need two local servers in the **same** job, use different **`container_name`** / **`llama_port`** inputs or run them in **separate** jobs.
-- **Runner cache layout:** The start action (`llama-server-docker-start`) takes **`llama_server_root`**. Under that directory it stores `model/model.gguf` and `docker/llama-server-image.tar` for `actions/cache`. This repo’s **integration** workflow uses **`${{ github.workspace }}/.cache/auto-pr-llama-stub`** / **`…-model`** so paths stay under the checkout (nested **`docker -v`** from [act](https://github.com/nektos/act) matches the host). Each integration job picks an **ephemeral TCP port** on the runner via [`scripts/integration-ephemeral-port.sh`](../scripts/integration-ephemeral-port.sh) (**`python3`**: `bind(127.0.0.1, 0)` — Python is preinstalled on GitHub-hosted Ubuntu; not inside nested containers). The **generate** reusable workflow still uses **`${{ runner.temp }}/auto-pr-llama`** for hosted runs.
+- **Runner cache layout:** The start action (`llama-server-docker-start`) takes **`llama_server_root`**. Under that directory it stores `model/model.gguf` and `docker/llama-server-image.tar` for `actions/cache`. This repo’s **integration** workflow uses **`${{ github.workspace }}/.cache/auto-pr-llama-stub`** / **`…-model`** so paths stay under the checkout (nested **`docker -v`** from [act](https://github.com/nektos/act) matches the host). Each integration job picks an **ephemeral TCP port** on the runner via an inline **`python3`** one-liner in [integration.yml](../.github/workflows/integration.yml) (`bind(127.0.0.1, 0)` — Python is preinstalled on GitHub-hosted Ubuntu; not inside nested containers). The **generate** reusable workflow still uses **`${{ runner.temp }}/auto-pr-llama`** for hosted runs.
 
 ## AI providers (`local`, `github-models`)
